perf: cache /country/{country_code} responses at startup
Precompute a HashMap<country_code, CountryResponse> in a OnceCell
during startup instead of scanning both MMDB databases on every
request. The previous handler walked 0.0.0.0/0 and ::0/0 via
within(...) and decoded every entry to filter by country code, which
made each request O(N) over the full dataset.

Changes:
- Add COUNTRY_CACHE static and init_country_cache() in db.rs that
  builds the map from a single pass over IPV4_COUNTRY and IPV6_COUNTRY,
  merging networks per country via entry().and_modify()/or_insert_with()
- Call init_country_cache() after init_mmdb() in main.rs
- Rewrite endpoint_get_country to do an O(1) HashMap lookup and
  return a cloned CountryResponse, dropping the per-request scan
- Derive Clone on Country and CountryResponse so cached values can
  be served without re-decoding
- Initialize the cache in the country_valid test
- Remove stale TODO comments about caching at startup
